Fix chart bug by multi lang support for date format
